+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2023-10-04
+
+### Fixed
+
+- Correctly handle single default exports, multiple named exports, and files with both default and named exports:
+  - Single default exports use `export * from './file';`
+  - Named exports use `export { namedExport1, namedExport2 } from './file'`
+  - Both default and named exports use `export { default as DefaultExport, namedExport1, namedExport2 } from './file';`
+- Added functionality to ignore specified files during indexing.
+
 ## [1.0.4] - 2023-10-04
+
+### Added
+
+- Added an `ignoreFiles` option in `config.json` to specify files that should be ignored during indexing.
+- By default, ignored `app/page.tsx` if `app` is in the watched folders.
 
 ### Fixed